test: consolidate suite — shared fakes, lint scripts, merged modules
Extract fake_pygame, merge touch-browser and SL HUD test files, move
systemd/ALSA string guards into offline lint scripts, trim duplicate
audio_engine parity tests, and wire repo lints into mpe test all.
