+++ tests/test_gesture_router.py
@@ -7,31 +7,11 @@ Documents/specs/touch-browser-browse-carousel-spec.md).
 
 from __future__ import annotations
 
-import sys
-import types
 import unittest
 
+from tests.fake_pygame import install_fake_pygame
 
-class _FakePygameRect:
-    def __init__(self, x, y, w, h):
-        self.x, self.y, self.w, self.h = x, y, w, h
-
-
-def _install_fake_pygame() -> None:
-    if isinstance(sys.modules.get("pygame"), types.ModuleType):
-        return
-    fake = types.ModuleType("pygame")
-    fake.error = type("error", (Exception,), {})
-    # patch_browser.geometry does `import pygame` once and the binding is
-    # cached for the process — a later, fuller pygame stub installed by
-    # another test module never reaches it. Since this file may be the
-    # first to import geometry (alphabetically early), the stub must
-    # cover geometry.Rect.pygame_rect's only dependency: Rect itself.
-    fake.Rect = _FakePygameRect
-    sys.modules["pygame"] = fake
-
-
-_install_fake_pygame()
+install_fake_pygame()
 
 from patch_browser.geometry import Rect  # noqa: E402
 from patch_browser.gesture_router import (  # noqa: E402
